move useUpdateShareMemberPermissions()
